appinfo
Write the title and appid of the currently active client to the
filesystem. This information can then be used by user scripts.

In order for the title and appid to be written to the filesystem the
dwl-getappinfo script must be provided as a startup script:

	dwl -s dwl-getappinfo

To use this script in conjunction with a status bar such as somebar use
the pipe tee utility `pee`:

	dwl -s 'pee somebar dwl-getappinfo'